ECMA-402 v1 legacy constructor semantics compromise for NumberFormat
This patch addresses #57 by allowing certain legacy constructor
patterns to coexist with the new guarantees in ECMA-402 v2, which
allows for a pattern where all internal slots exist from the
beginning of the object's lifetime. The compromise is based on
storing a "real" object inside of a symbol-named property to allow
for object initialization in cases of the
Intl.<constructor>.call(Object.create(Intl.<constructor>) pattern.
Legacy methods have to forward their calls to this "real" object.

This patch specifies the change for Intl.NumberFormat, and a follow-on
patch makes the same change for Intl.DateTimeFormat.

This patch, together with changes for Intl.DateTimeFormat,
has been demonstrated to fix old versions of Intl.js on a deployed
website.

A sample implementation in V8 can be found at
https://codereview.chromium.org/1828543007