> https://tc39.github.io/ecma402/#annex-implementation-dependent-behaviour needs to be updated, too.
added. "List of Strings representing the possible results of plural selection and their corresponding order per locale." seems to be the only implementation dependent-behavior for PluralRules

> Also allow higher values in https://tc39.github.io/ecma402/#conformance ?

I have added: "The options property `type` in the PluralRule constructor." assuming that there might be some other types that might be implementation-defined. But I'm not sure if that's what you're referencing to with "higher values". Can you explain?

> Add PluralRules to https://tc39.github.io/ecma402/#locale-and-parameter-negotiation ?

Added there for the list that supports `[[LocaleData]]`, and fixed `InitializePluralRules` algo to call `ResolveLocale` with 5 arguments.