Normative: Improve handling of non-Gregorian calendars
In CLDR, the set of patterns is based on the the calendar, not
simply the language-region-script. Previously, however, the set
was based purely on the base name. This patch switches to a two-level
lookup to permit implementations to have more accuracy.

The patch adds an additional field to DateTimeFormat: relatedYear.
This is the related Gregorian year, e.g., as used in some Chinese calendar
formats when the Gregorial year is put next to the year in a 60-year cycle.

The patch makes an editorial change to make the descriptions
of the locale database fully switch to records (there was some object-style
notation remaining).

Closes #225