Normative: Apply kebab-case for best-fit
This is an initial effort to apply kebab-case for enumeration string values in ECMA-402.

This style follow convention set at the W3C Tag Design Principles. Ref:
https://w3ctag.github.io/design-principles/#casing-rules

The goal is to illustrate a change in an experimental way, it first needs discussion from
the Intl team to confirm if 'best fit' should also be included in a possible set of changes
to match style convention.