Editorial: Add type assertion to GetOption
In Temporal, we are using an operation similar to GetOption (and intend
to add it to 262 in the future) and it came up in a discussion that it
would be better to have this assertion in GetOption to show that all
call sites should perform ToObject or whatever on the options object.

Assertion also added to GetNumberOption for symmetry, and to
ApplyOptionsToTag for clarity.