Editorial: Add GetOptionsObject (#480)
In order to later be able to move GetOption into ECMA-262 in sync with
Temporal, we add GetOptionsObject and CoerceOptionsToObject as abstract
operations.

GetOptionsObject and GetOption are intended to be moved into ECMA-262 at
the same time, while CoerceOptionsToObject is intended to stay in
ECMA-402. CoerceOptionsToObject describes the somewhat unexpected boxing
behaviour when passing a primitive as the options parameter in a formatter
constructor. This behaviour may need to be preserved for web-compatibility
reasons, but new Intl objects going forward should use the
GetOptionsObject behaviour.