Editorial: Fix linter errors in spec
Based on a discussion in Temporal [1], it seems like it might be a good
idea to run ecmarkup's linter on the spec, now that in version 4.1.0 has
brought several bug fixes in the linter.

This adds the linter to the build process, but without making it fail
the build if any warnings are detected, since some of the "errors" are
intentional. (The intentional ones are marked with a comment.)

[1] https://github.com/tc39/proposal-temporal/pull/870