Editorial: Refactor time zone calculation in ToLocalTime
The operation LocalTZA from ECMA-262 has been replaced by other, more
clearly specified, mechanisms for determining the time zone offset.

(The language "LocalTZA() where the local time zone is replaced with
_timeZone_" seemed insufficiently specific, anyway.)

This uses the new operation GetNamedTimeZoneOffsetNanoseconds from
ECMA-262 to calculate the time zone offset. This operation encapsulates
the requirement to "use best available information about the specified
time zone, including current and historical information about time zone
offsets from UTC and daylight saving time rules," so we can delete that
text from the non-Gregorian branch in this operation.

Add a structured header while we're at it.

In addition, I believe the types of _t_ and _tz_ were backwards. _t_ is an
exact time in milliseconds, so it is a time value. Elsewhere we ensure _t_
is not NaN, so we can say it is a finite time value. _tz_ may not be a
time value once _offsetMs_ (formerly _timeZoneOffset_) is added to it, so
it is an integral Number.