Editorial: Change ToLocalTime to use epoch nanoseconds
This is an optional third commit that changes ToLocalTime to use epoch
nanoseconds instead of milliseconds, in anticipation of Temporal becoming
part of the specification. However, there is a small advantage to bringing
this change in now, because using epoch nanoseconds fits better with the
adaptation to use GetNamedTimeZoneOffsetNanoseconds from the previous
commit, which also takes epoch nanoseconds.

I've taken the liberty of changing the "Operation(x) specified in es2023's
Operation" language to just use the automatic linking through the
ecma262 bibliography.