Normative: Canonicalise "GMT" to "UTC"
tzdata 2022f using vanguard format made "GMT" a zone and "Etc/GMT" a
link to "GMT". Rearguard format still has this the other way around,
i.e. "Etc/GMT" is a zone and "GMT" a link to "Etc/GMT". Because we don't
control if vanguard or rearguard format is used by an implementation, the
`CanonicalizeTimeZoneName` operation should canonicalise both "Etc/GMT"
and "GMT" to "UTC".