editorial: CreatePartsFromList type change
Change the expected type of the "list" argument for the
CreatePartsFromList abstract operation from a List of String values to a
simple List. This makes no outside observable normative change but
allows other built-ins to use this operation in more interesting ways,
primarily motivated by `Intl.DurationFormat` which utilizes this
operation in order to format the list of different duration units.

Refs: https://github.com/tc39/proposal-intl-duration-format/issues/103