Editorial: Set default value for fractionalSecondDigits in FormatDateTimePattern
When `FormatDateTimePattern` is called from
`PartitionDateTimeRangePattern` to format a range pattern, the range
pattern could contain the substring `"fractionalSecondDigits"`, but the
DateTimeFormat's `[[FractionalSecondDigits]]` internal slot is
`undefined`. This makes operations like `10^(fractionalSecondDigits-3)`
invalid. (This also means this change has to be editorial, because it's
an internal spec bug.)

It seems difficult to describe a restriction that range patterns mustn't
include smaller calendar elements when compared to the corresponding
non-range DateTime Format Record. Therefore this commit simply defaults
`fractionalSecondDigits` to a default value when
`[[FractionalSecondDigits]]` is `undefined`.

This merely theoretical case doesn't happen in actual implementations,
so implementors can ignore this change.