Editorial: Use correct field access in PartitionDateTimeRangePattern
The "Range Pattern Field" column is of type `Field Name`, but the
`[[<name>]]` form can only be used when `name` is a `String` value.
Either the "Range Pattern Field" column values need to be changed to
String values, e.g. `"Era"` instead of `[[Era]]`, or the field access
needs to be rewritten. This commit uses the latter approach.