Editorial: Correct 2-digit handling in FormatDateTimePattern
Two issues:
1. `"length" property of fv` isn't the correct term to determine a
   string length.
2. And it doesn't correctly handle digits which contain surrogate pairs.

Use `StringToCodePoints` to correctly determine the number of code
points and extract the last two code points. And then call
`CodePointsToString` to compute the resulting string.

`CodePointsToString` is declared as receiving a `sequence of Unicode
code points`, not a `List of Unicode code points`, but it seems like the
`sequence` and `List` types are interchangeable, at least based on other
uses in the ECMA-262 spec.