Editorial: Replace AvailableCanonicalCalendars with AvailableCalendars
AvailableCalendars should return all possible aliases, so that other
places in the spec (e.g. in the future, validating a string calendar ID in
Temporal) can use them to determine whether a given input value is valid.
This input value can subsequently be canonicalized by another abstract
operation, CanonicalizeUValue, which we can use in several other places.

In Intl.supportedValuesOf(), on the other hand, we should not return all
possible aliases, so we filter them out using CanonicalizeUValue before
returning the list of AvailableCalendars codes as an array to the caller.

See https://github.com/tc39/proposal-intl-enumeration/pull/49. This is the
part of that PR that I consider relevant for the future integration of
Temporal. The time zone parts were already done as part of #876. If
desired, I could implement the rest of that PR, additionally supporting
collation, currency, numbering system, and unit canonicalization.

Closes: #726