Normative: Don't add default formatting to lone era
Previously, if `era` was given by itself in the formatting options, the
default formatting for the other date/time components would be added.
This is inconsistent with what we do for `dayPeriod` where a lone
`dayPeriod` produces formats such as "in the afternoon".

The use case of obtaining the localized era name has also been mentioned
several times.

This change makes a lone `era` be passed to the format matcher algorithm
without added date/time components, and requires implementations to have
at least a year-month-day-era and year-month-era format available (which
looks to be widely available across languages in CLDR.)

See: #461