Normative: Don't add default formatting to lone timeZoneName
Previously, if `timeZoneName` was given by itself in the formatting
options, the default formatting for the other date/time components would
be added. This is inconsistent with what we do for `dayPeriod` where a
lone `dayPeriod` produces formats such as "in the afternoon".

The use case of obtaining the localized time zone name in all its
possible variants such as specific/generic, has been mentioned several
times.

This change makes a lone `timeZoneName` be passed to the format matcher
algorithm without added date/time components, and requires
implementations to have at least a
year-month-day-hour-minute-second-timeZoneName format available.

See: #461